# backend/docs/swagger.yaml
definitions:
...
  custom_errors.HTTPError:
    properties:
      error_description:
        type: string
    type: object
  enums.Color:
    enum:
    - Red
    - Blue
    - Yellow
    - Green
    - Orange
    - Purple
    - Brown
    - Black
    - White
    - Gray
    - Cyan
    - Magenta
    - Lime
    - Navy
    - Teal
    - Maroon
    - Olive
    - Beige
    - Gold
    - Other
    type: string
    x-enum-varnames:
    - RED
    - BLUE
    - YELLOW
    - GREEN
    - ORANGE
    - PURPLE
    - BROWN
    - BLACK
    - WHITE
    - GRAY
    - CYAN
    - MAGENTA
    - LIME
    - NAVY
    - TEAL
    - MAROON
    - OLIVE
    - BEIGE
    - GOLD
    - OTHER
  enums.Drive:
    enum:
    - FWD
    - RWD
    - AWD
    type: string
    x-enum-varnames:
    - FWD
    - RWD
    - AWD
  enums.FuelType:
    enum:
    - Diesel
    - Petrol
    - Electric
    - Ethanol
    - LPG
    - Biofuel
    - Hybrid
    - Hydrogen
    type: string
    x-enum-varnames:
    - DIESEL
    - PETROL
    - ELECTRIC
    - ETHANOL
    - LPG
    - BIOFUEL
    - HYBRID
    - HYDROGEN
  enums.MarginValue:
    enum:
    - 3
    - 5
    - 10
    type: integer
    x-enum-varnames:
    - LOW_MARGIN
    - MEDIUM_MARGIN
    - HIGH_MARGIN
  enums.Transmission:
    enum:
    - Manual
    - Automatic
    - Cvt
    - Dual clutch
    type: string
    x-enum-varnames:
    - MANUAL
    - AUTOMATIC
    - CVT
    - DUAL_CLUTCH
...
  pagination.PaginationRequest:
    properties:
      page:
        type: integer
      page_size:
        type: integer
    type: object
  pagination.PaginationResponse:
    properties:
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  sale_offer.CreateSaleOfferDTO:
    properties:
      color:
        $ref: '#/definitions/enums.Color'
      description:
        type: string
      drive:
        $ref: '#/definitions/enums.Drive'
      engine_capacity:
        type: integer
      engine_power:
        type: integer
      fuel_type:
        $ref: '#/definitions/enums.FuelType'
      manufacturer:
        type: string
      margin:
        $ref: '#/definitions/enums.MarginValue'
      mileage:
        type: integer
      model:
        type: string
      number_of_doors:
        type: integer
      number_of_gears:
        type: integer
      number_of_seats:
        type: integer
      price:
        type: integer
      production_year:
        type: integer
      registration_date:
        type: string
      registration_number:
        type: string
      transmission:
        $ref: '#/definitions/enums.Transmission'
      vin:
        type: string
    required:
    - color
    - description
    - drive
    - engine_capacity
    - engine_power
    - fuel_type
    - manufacturer
    - margin
    - mileage
    - model
    - number_of_doors
    - number_of_gears
    - number_of_seats
    - price
    - production_year
    - registration_date
    - registration_number
    - transmission
    - vin
    type: object
  sale_offer.MinMax-string:
    properties:
      max:
        type: string
      min:
        type: string
    type: object
  sale_offer.MinMax-uint:
    properties:
      max:
        type: integer
      min:
        type: integer
    type: object
  sale_offer.OfferFilter:
    properties:
      car_registration_date_range:
        $ref: '#/definitions/sale_offer.MinMax-string'
      colors:
        items:
          $ref: '#/definitions/enums.Color'
        type: array
      drives:
        items:
          $ref: '#/definitions/enums.Drive'
        type: array
      engine_capacity_range:
        $ref: '#/definitions/sale_offer.MinMax-uint'
      engine_power_range:
        $ref: '#/definitions/sale_offer.MinMax-uint'
      fuel_types:
        items:
          $ref: '#/definitions/enums.FuelType'
        type: array
      is_order_desc:
        type: boolean
      liked_only:
        type: boolean
      manufacturers:
        items:
          type: string
        type: array
      mileage_range:
        $ref: '#/definitions/sale_offer.MinMax-uint'
      offer_creation_date_range:
        $ref: '#/definitions/sale_offer.MinMax-string'
      offer_type:
        $ref: '#/definitions/sale_offer.OfferType'
      order_key:
        type: string
      price_range:
        $ref: '#/definitions/sale_offer.MinMax-uint'
      query:
        type: string
      transmissions:
        items:
          $ref: '#/definitions/enums.Transmission'
        type: array
      user_id:
        type: integer
      year_range:
        $ref: '#/definitions/sale_offer.MinMax-uint'
    type: object
  sale_offer.OfferFilterRequest:
    properties:
      filter:
        $ref: '#/definitions/sale_offer.OfferFilter'
      pagination:
        $ref: '#/definitions/pagination.PaginationRequest'
    type: object
  sale_offer.OfferType:
    enum:
    - Regular offer
    - Auction
    - Both
    type: string
    x-enum-varnames:
    - REGULAR_OFFER
    - AUCTION
    - BOTH
  sale_offer.RetrieveDetailedSaleOfferDTO:
    properties:
      brand:
        type: string
      buy_now_price:
        type: integer
      can_modify:
        type: boolean
      color:
        $ref: '#/definitions/enums.Color'
      date_end:
        type: string
      date_of_issue:
        type: string
      description:
        type: string
      drive:
        $ref: '#/definitions/enums.Drive'
      engine_capacity:
        type: integer
      engine_power:
        type: integer
      fuel_type:
        $ref: '#/definitions/enums.FuelType'
      id:
        type: integer
      images_urls:
        items:
          type: string
        type: array
      is_auction:
        type: boolean
      is_liked:
        type: boolean
      margin:
        $ref: '#/definitions/enums.MarginValue'
      mileage:
        type: integer
      model:
        type: string
      number_of_doors:
        type: integer
      number_of_gears:
        type: integer
      number_of_seats:
        type: integer
      price:
        type: integer
      production_year:
        type: integer
      registration_date:
        type: string
      registration_number:
        type: string
      status:
        type: string
      transmission:
        $ref: '#/definitions/enums.Transmission'
      user_id:
        type: integer
      username:
        type: string
      vin:
        type: string
    type: object
  sale_offer.RetrieveOffersWithPagination:
    properties:
      offers:
        items:
          $ref: '#/definitions/sale_offer.RetrieveSaleOfferDTO'
        type: array
      pagination:
        $ref: '#/definitions/pagination.PaginationResponse'
    type: object
  sale_offer.RetrieveSaleOfferDTO:
    properties:
      can_modify:
        type: boolean
      color:
        $ref: '#/definitions/enums.Color'
      id:
        type: integer
      is_auction:
        type: boolean
      is_liked:
        type: boolean
      main_url:
        type: string
      mileage:
        type: integer
      name:
        type: string
      price:
        type: integer
      production_year:
        type: integer
      status:
        type: string
      username:
        type: string
    type: object
  sale_offer.UpdateSaleOfferDTO:
    properties:
      color:
        $ref: '#/definitions/enums.Color'
      description:
        type: string
      drive:
        $ref: '#/definitions/enums.Drive'
      engine_capacity:
        type: integer
      engine_power:
        type: integer
      fuel_type:
        $ref: '#/definitions/enums.FuelType'
      id:
        type: integer
      manufacturer:
        type: string
      margin:
        $ref: '#/definitions/enums.MarginValue'
      mileage:
        type: integer
      model:
        type: string
      number_of_doors:
        type: integer
      number_of_gears:
        type: integer
      number_of_seats:
        type: integer
      price:
        type: integer
      production_year:
        type: integer
      registration_date:
        type: string
      registration_number:
        type: string
      transmission:
        $ref: '#/definitions/enums.Transmission'
      vin:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: BD2
  description: Car-Dealer API
  license:
    name: MIT
  title: Car‑Dealer API
  version: "1.0"
paths:
...
  /sale-offer:
    post:
      consumes:
      - application/json
      description: |-
        Creates a new sale offer in the database. To create a sale offer, the user must be logged in. There are several constraints on the offer fields, such as:
        - Color must be one of the predefined colors (endpoint: /car/colors)
        - Fuel type must be one of the predefined fuel types (endpoint: /car/fuel_types)
        - Transmission must be one of the predefined transmission types (endpoint: /car/transmissions)
        - Drive must be one of the predefined drive types (endpoint: /car/drives)
        - Model must be one of the predefined models (endpoint: /car/models or /car/models/:id)
        - Number of doors must be between 1 and 6
        - Number of seats must be between 2 and 100
        - Engine power must be less than or equal to 9999 (in horsepower)
        - Engine capacity must be less than or equal to 9000 (in cm3)
        - Number of gears must be between 1 and 10
      parameters:
      - description: Sale offer form
        in: body
        name: offer
        required: true
        schema:
          $ref: '#/definitions/sale_offer.CreateSaleOfferDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created - returns the created sale offer
          schema:
            $ref: '#/definitions/sale_offer.RetrieveDetailedSaleOfferDTO'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized - user not logged in
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - Bearer: []
      summary: Create a new sale offer
      tags:
      - sale-offer
    put:
      consumes:
      - application/json
      description: Updates an existing sale offer in the database. To update a sale
        offer, the user must be logged in and must be the owner of the offer. Constraints
        are the same as when creating a sale offer.
      parameters:
      - description: Sale offer form
        in: body
        name: offer
        required: true
        schema:
          $ref: '#/definitions/sale_offer.UpdateSaleOfferDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated - returns the updated sale offer
          schema:
            $ref: '#/definitions/sale_offer.RetrieveDetailedSaleOfferDTO'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized - user must be logged in to update his offer
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "403":
          description: Forbidden - user can only update his own offer
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "404":
          description: Sale offer not found
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - Bearer: []
      summary: Update a sale offer
      tags:
      - sale-offer
  /sale-offer/buy/{id}:
    post:
      consumes:
      - application/json
      description: Allows a user to buy an item from a sale offer
      parameters:
      - description: Sale Offer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully purchased offer
        "401":
          description: Unauthorized - user must be logged in to buy an offer
        "403":
          description: Forbidden - user cannot buy his own offer
        "404":
          description: Not Found - sale offer not found
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Buy a sale offer
      tags:
      - sale-offer
  /sale-offer/filtered:
    post:
      consumes:
      - application/json
      description: |-
        Returns a list of sale offers in paginated form. If the user is logged in, the results contain he offers created by the user. The results are filtered based on request's body. There are several constraints on the filter fields, such as:
        - Auction type must be one of the predefined offer types (endpoint: /sale-offer/offer-types)
        - Order key must be one of the predefined order keys (endpoint: /sale-offer/order-keys)
        - List of manufacturers must contain only predefined manufacturers (endpoint: /car/manufacturers)
        - List of colors must contain only predefined colors (endpoint: /car/colors)
        - List of drives must contain only predefined drives (endpoint: /car/drives)
        - List of fuel types must contain only predefined fuel types (endpoint: /car/fuel_types)
        - List of transmissions must contain only predefined transmission types (endpoint: /car/transmissions)
        - Whenever you use a range, the min value must be less than or equal to the max value, you can provide only one of them, and the other will be ignored.
      parameters:
      - description: Sale offer filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/sale_offer.OfferFilterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of sale offers
          schema:
            $ref: '#/definitions/sale_offer.RetrieveOffersWithPagination'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get filtered sale offers
      tags:
      - sale-offer
  /sale-offer/id/{id}:
    get:
      consumes:
      - application/json
      description: Returns a sale offer by its ID. Can be used to retrieve detailed
        information about sale offer.
      parameters:
      - description: Sale offer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sale offer details
          schema:
            $ref: '#/definitions/sale_offer.RetrieveDetailedSaleOfferDTO'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "404":
          description: Sale offer not found
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get sale offer by ID
      tags:
      - sale-offer
  /sale-offer/my-offers:
    post:
      consumes:
      - application/json
      description: Returns a list of all sale offers created by the logged-in user.
      parameters:
      - description: Pagination request
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/pagination.PaginationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of sale offers
          schema:
            $ref: '#/definitions/sale_offer.RetrieveOffersWithPagination'
        "401":
          description: Unauthorized - user must be logged in to retrieve his offers
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - Bearer: []
      summary: Get my sale offers
      tags:
      - sale-offer
...
schemes:
- http
swagger: "2.0"